Update patch set 1
Patch Set 1:

The arrays are probably in the BSS segment and probably do not affect the size of the program on disk.  But some people do run Go in relatively memory constrained environments.  I wouldn't worry if this were widely useful, but in practice very few people are going to ever see this information.

For what it's worth, a core dump should include the heap as well as the stack.

Patch-set: 1