Update patch set 4
Patch Set 4:

you mean the code changes should be in go/build?
I thought we wanted to make this change only in the tooling, leaving the stdlib as is.

would you rather have it as a change on the defaultContext() function in go/build
that's also an easy change

Patch-set: 4
Label: Code-Review=0