Update patch set 1
Patch Set 1:

(5 comments)

The more I looked at this CL the more I agree with Brad that this code really can't just be patched up more. The structure needs to be made clearer. I sent CL 32092 to do that. Once that's in, we can make this a test-only CL.

Patch-set: 1