Update patch set 9
Patch Set 9:

> > If ~/go does not exist, GOPATH is not set, and we do go get X the
 > > directory is created.
 > > This behavior is the one I would expect, as frictionless as
 > > possible.
 > 
 > I concur. This is what happens now
 > 
 > env GOPATH=$(mktemp -d) go get github.com/pkg/profile
 > 
 > will create $GOPATH/go/github.com/pkg/profile, which nobody seems
 > to complain about.
 
But this is a different example. When you run

env GOPATH=$(mktemp -d) ...

the mktemp created $GOPATH. Arguably even if you run

env GOPATH=/tmp/notexist ...

you're explicitly asking for that GOPATH, so I don't mind if the 'go get' creates /tmp/notexist (and I believe it does, although that surprised me when I realized it a few days ago).

My question is whether, with an unset $GOPATH, 'go get' should create $HOME/go (the default $GOPATH). That might be a bit too much. We might want to require the user to either set $GOPATH or create $HOME/go, to get some signal that they want to use that directory. If we did that, then we'd want to change the appropriate error to something like "GOPATH not set and $HOME/go does not exist".

Let's do that - avoid setting $GOPATH in go/build when the thing you want to set it to does not exist yet - and see how far that takes us. We can always decide to force the creation of the default directory later, but it's harder to undo later.

Patch-set: 9