Update patch set 2
Patch Set 2:

When GC begins all of the mark workers are parked and have a very well defined, very shallow stack. If we arrange to have no pointers that need to be scanned we have what amounts to an empty stack as far as scanning is concerned. Why can't we just trivially declare them scanned by construction?

Patch-set: 2
Label: Code-Review=0