Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> When GC begins all of the mark workers are parked and have a very well defined, very shallow stack. If we arrange to have no pointers that need to be scanned we have what amounts to an empty stack as far as scanning is concerned. Why can't we just trivially declare them scanned by construction?

This was my first approach, but it turns out not to be true. If a mark worker completes the GC cycle, it can end up with a non-trivial stack where we don't have much control over whether or not there are heap pointers. In fact, there's at least one known heap pointer: if two mark workers race to complete the GC cycle, the one that blocks on markDoneSema can heap allocate a sudog to perform the wait and then go to sleep.

That's why I switched to this new approach where we don't have to be careful about what's on the stack for the whole breadth and depth of the mark worker. With this approach we just need to be careful across a handful of relatively simple lines (and not recursively).

Patch-set: 2