Update patch set 2
Patch Set 2:

> Patch Set 1: Code-Review-1
> 
> I don't think the new version is more readable
> than the previous switch based version.
> 
> I'd rather make the compiler do the correct
> thing than to rewrite switches into one line
> boolean expressions.

Compromise: Joe, file a bug about the compiler shortcoming and drop a TODO in the code to consider reverting to the switch in the future when the compiler is smarter.

But I personally don't think the new single-line code is less readable (or at least not much less readable) enough to warrant not getting the performance.

Patch-set: 2