Update patch set 1
Patch Set 1:

> Ah, you're right, Martin. When I bump the []byte size, it starts
 > allocating.
 > 
 > https://play.golang.org/p/TjraQaD2dm

we stack allocate the backing array up to 32byte.

I had tried a CL to optimize range string([]byte) but its hard to verify it is safe since the code in the range can not be allowed to modify the []byte array. So no function calls or changes of values that could be in the backing array which does not leave many range statements this could be applied to.

Patch-set: 1