Update patch set 10
Patch Set 10:

> > Personally I think that automatically creating the ~/go directory
 > > is the right thing to do.  Perhaps with a message saying that we
 > > are doing it.
 > 
 > Yes, that's my take on it also.
 > 
 > I propose:
 > 
 > - If $GOPATH is unset
 > - default to $HOME/go.
 > - if $HOME/go does not exist
 > - create it
 > - print a message informing the user
 > - if $HOME/go exists
 > - if it contains a src directory, use it as GOPATH
 > - if it does not contain a 'src' directory, exit with an error
 > "$HOME/go exists but does not look like a $GOPATH"
 > (or similar)
 > - If $GOPATH is set
 > - create and use the specified $GOPATH

OK, but if we're going to create $HOME/go when it doesn't exist, then 'mkdir go' by itself shouldn't break things because no there's a $HOME/go but no $HOME/go/src. The go command already does all the mkdirs it needs to do. It sounds like maybe we should print a message saying that we created GOPATH (no matter whether it was from the default), but otherwise things are good.

Patch-set: 10