Update patch set 10
Patch Set 10:

> > Those are not exported in the spec sense (no capital letter) so
 > > they cannot be part of a qualified identifier; that is the answer
 > > is no.
 > 
 > Sorry, I meant can I declare them as aliases, not can they be
 > aliased.  For example:
 > 
 > func main => foo.Main
 > func init => bar.Init
 > func init => quux.Init

Duh, sorry for being dim.

No, this should not be permitted (at least 'init' should be excluded). The compiler accepts main as alias right now but the linker fails. An init alias leads to a redeclaration error.

Since we've gone though the review cycle with this and it's essentially errors that should be better (and the spec more accurate), in the interest of progress, let's leave it as is here and update in a future CL. I filed issue #17637.

Patch-set: 10
Label: Code-Review=0