Update patch set 3
Patch Set 3:

I'm not sure that we should get this in before the freeze I'm afraid:

Bigint / elliptic curve code is very prone to very hard to find bugs that, none the less, can be security critical if deliberately targeted.

Reviewing these implementations can take weeks of time, and that's for common architectures!

Patch-set: 3