Update patch set 3
Patch Set 3:

(5 comments)

As you might expect, neither Adam nor I can reliably review 2,400+ lines of assembly for an unfamiliar architecture. I will be happy to review the way it's fitting into the rest of the package (comments below), but we are going to have to rely on the Go on s390x team at IBM for the review of the actual assembly.

Michael Munday, can you please review for the assembly?

Thanks.
Russ

Patch-set: 3