cmd/compile: do more type conversion inline
The code to do the conversion is smaller than the
call to the runtime.
The asserts need to call panic if they fail, but that
code is out of line.

The only conversions left in the runtime are those which
might allocate (convT2? for non-direct T) and those which
ask for an itab (assert?2I?).

DO NOT REVIEW
DO NOT SUBMIT

Fixes #17405
Fixes #8422

Change-Id: Ida2367bf8ce3cd2c6bb599a1814f1d275afabe21