Update patch set 2
Patch Set 2: Code-Review+2

This seems to do what it says on the tin, but are we not worried about the tests we will break with this? I have to imagine some users are currently (ab)using this feature.

Patch-set: 2
Reviewer: Gerrit User 13020 <13020@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=+2