Update patch set 4
Patch Set 4:

This makes me a bit nervous—I'd like to know why the nilling out was there to begin with and why it's ok to remove now. Want to do a bit of git blame and find out? If not, it's ok, I'll dig myself in a few days.

Patch-set: 4