Update patch set 4
Patch Set 4:

Sure. Origin of the nilling out is [1]. That is when type-checking was initially implemented in C and parsers were in YACC. It's been there ever since, virtually unmodified. May be we should we ask rsc@ (he originally added that logic in 2009).

Full history: [A] -> [B] -> [C] -> [D], Where [A] is most recent and [D] is oldest.

[1]: https://github.com/golang/go/commit/9dc22b6d6fbd62ce127dbad1bdbfbb073ad7c4af#diff-f360edfb3352ea99525e85e2220369a9R1538

[A]: https://github.com/golang/go/blame/833f57ed502d686ce054ff557cf3bf5b6ed1a1d2/src/cmd/compile/internal/gc/typecheck.go#L3116

[B]: https://github.com/golang/go/blame/d0b59deb71b69c1a07b3e734b6f5ff66b3a4e2c8/src/cmd/internal/gc/typecheck.go#L3078

[C]: https://github.com/golang/go/blame/c11882bc3e8f6ab4eea79d7bbd0647b31dad7482/src/cmd/gc/typecheck.c#L2785

[D]: https://github.com/golang/go/blame/178089056ea9dae2c2807f0b5b475c2caf09fde3/src/cmd/gc/typecheck.c#L383

Patch-set: 4