Update patch set 1
Patch Set 1:

> The arrays are probably in the BSS segment and probably do not affect the size of the program on disk.

Thanks for confirming.

 > But some people do run Go in relatively memory constrained environments.

Is virtual memory also generally constrained in those environments? Would it be helpful to save panic messages in an array/slice that would only get paged in as the process crashed?

 > I wouldn't worry if this were widely useful, but in practice very few people are going to ever see this information.

Yes, I agree. Thank you for making clear how this affects the budget for overhead.

 > For what it's worth, a core dump should include the heap as well as the stack.

Yes, though I'm not sure when to allocate a slice to store the panic text.

- Allocating during a panic/fatal doesn't seem like a practical option, since the crash may be due to heap corruption. This could turn a well-behaved debuggable "fatal error" into a segfault. It also means that we'd miss any text leading up to "fatal error"s, such as the detailed descriptions of weird stack contents.

- Gating the allocation by GOTRACEBACK=crash might work, though I'm a bit uncomfortable with changing the behavior of this code based on that setting. If a user has an application that crashes sometimes and they enable core dumps in order to debug, it would be unfortunate for the debug setting to lead to additional problems.

I'm not sure how to proceed with this change.

Maybe part of the solution to https://golang.org/issue/16410 ("create a heap dump viewer") could be to print the output that runtime.printpanics/runtime.dopanic_m would have printed during the crash. Then the only thing a patch like this one would need to do is to retain the pre-panic output from "fatal error" messages..

Can you offer advice on how—or if—to proceed?

Patch-set: 1
Label: Code-Review=0