Update patch set 2
Patch Set 2: -Run-TryBot

> (1 comment)
 > 
 > Makes sense to me I guess. Would anything really bad happen if we
 > stopped lying about the ptr-ness of moduledata? I guess the GC
 > would spend a bunch of time pointless scanning static data...

I don't have a good sense of how much more work that would be for the GC. I did consider an entirely new tracking object, but it's a bit of a pain without keeping a *moduledata in it, and it gets to feel a little heavy. Maybe if there were more heap pointers that needed tracking.

Patch-set: 2
Label: Run-TryBot=0