Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> As far as I see there is currently a race condition that can result in CPU underutilization. The race is between a P going to sleep and enlistWorker. Findrunnable checks gcMarkWorkAvailable only once before dropping P. As the result a P can check gcMarkWorkAvailable and observe that there is no work, then a GC goroutine executes putfull and calls enlistWorker, enlistWorker sees that npidle==0 and decides to not wake another P; the first P blocks. Findrunnable should recheck gcMarkWorkAvailable _after_ dropping P.

That's true, but I'm also not concerned about it. In fact, if we just enqueue one workbuf, I'm perfectly happy to not start an idle worker because we're probably going to handle that workbuf right away anyway and the idle wakeup would be a waste. The real problem is when we get into a steady state with idle Ps and in steady state we're going to be enqueuing workbufs very frequently.

I also don't want enlistWorkers to require expensive synchronization, since it's called on every buffer put.

Patch-set: 1