Update patch set 4
Patch Set 4:

> Anything else? I would like this to go in before the freeze, if
 > possible.

Do you have any follow on CLs? My only concern with this is that it is all dead code at the moment (you should address Lynn's earlier comment too).

If you do have a CL that uses one or more of these fields you can rebase it on top of this one and mail it before the freeze, even if this one hasn't been submitted.

Patch-set: 4