Update patch set 4
Patch Set 4:

> Do you have any follow on CLs? My only concern with this is that it
 > is all dead code at the moment (you should address Lynn's earlier
 > comment too).

Not at the moment. I have one change that will use it, but I'm afraid I cannot submit it today. 

Regarding Lynn's comment, how exactly a testcase would be implemented for this? Assuming the implementation is correct, I think the only way this could fail is if the kernel doesn't export the auxv correctly. Is there an alternative way of checking the processor so I can compare with the value in the cpu variable? 

 > 
 > If you do have a CL that uses one or more of these fields you can
 > rebase it on top of this one and mail it before the freeze, even if
 > this one hasn't been submitted.

Got it. Thanks for the clarification.

Patch-set: 4
Reviewer: Gerrit User 13015 <13015@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=0