cmd/link: use plugin path in deadcode analysis
CL 32355 (commit 9da7058) switched from using the output file as a
plugin prefix to the full package path. The linker dead code analysis
was not updated.

In addition, some very tiny plugins seem to suffer from overly
aggressive dead code analysis of types that are otherwise needed in
parts of the runtime. Be a bit more conservative about what from the
runtime we keep in a plugin.

Updates #17709 (and may fix it completely).

Change-Id: I01e50fb529791f3134f7aa294819931a11157c96