Update patch set 13
Patch Set 13:

(1 comment)

This CL has stopped being productive. The discussion on #17262 was about assuming GOPATH=$HOME/go when it is not set. That's a three-line change. 

This CL is mired in design discussions about when to accept a GOPATH and when not to, when to create it, and so on. A CL is not the place to have a design discussion. Let's go back to the issue and actually write down the design first.

Patch-set: 13